+++ packages/backend/src/core/PerpetualRollupSyncService.ts
@@ -9,7 +9,7 @@ import { PerpetualRollupStateTransitionCollector } from './collectors/PerpetualR
 import { UserRegistrationCollector } from './collectors/UserRegistrationCollector'
 import { UserTransactionCollector } from './collectors/UserTransactionCollector'
 import { VerifierCollector } from './collectors/VerifierCollector'
-import { WithdrawableAssetCollector } from './collectors/WithdrawableAssetsCollector'
+import { WithdrawableAssetCollector } from './collectors/WithdrawableAssetCollector'
 import { IDataSyncService } from './IDataSyncService'
 import { PerpetualRollupUpdater } from './PerpetualRollupUpdater'
